chore(uiDataSanitisation): split up functions, fix circular dependency issue
