+++ src/ui/server/helpers/sanitise-data/index.ts
@@ -1,36 +1,29 @@
 import { isValid as isValidDate } from 'date-fns';
-import { isNumber } from '../number';
-import { isEmptyString, stripCommas } from '../string';
+import { isEmptyString } from '../string';
 import { isAnObjectWithKeysAndValues } from '../object';
-import { FIELD_IDS } from '../../constants';
-import isValidWebsiteAddress from '../is-valid-website-address';
-import { SanitiseValueObjParams, RequestBody } from '../../../types';
+import sanitiseObject from './sanitise-object';
+import sanitiseArrayOfStrings from './sanitise-array-of-strings';
+import INSURANCE_FIELD_IDS from '../../constants/field-ids/insurance';
+import sanitiseValue from './sanitise-value';
+export * from './sanitise-value';
+import { RequestBody } from '../../../types';
 
 type ObjectType = {
   [key: string]: any;
 };
 
 const {
-  ACCOUNT: { SECURITY_CODE },
   EXPORTER_BUSINESS: {
-    COMPANY_HOUSE: { COMPANY_NUMBER, COMPANY_SIC },
-    YOUR_COMPANY: { PHONE_NUMBER },
-    NATURE_OF_YOUR_BUSINESS: { GOODS_OR_SERVICES, YEARS_EXPORTING, EMPLOYEES_INTERNATIONAL, EMPLOYEES_UK },
+    NATURE_OF_YOUR_BUSINESS: { YEARS_EXPORTING, EMPLOYEES_INTERNATIONAL, EMPLOYEES_UK },
     TURNOVER: { ESTIMATED_ANNUAL_TURNOVER, PERCENTAGE_TURNOVER },
-    BROKER: { ADDRESS_LINE_1, ADDRESS_LINE_2, POSTCODE, NAME: BROKER_NAME },
   },
   POLICY: {
     CONTRACT_POLICY: {
-      CREDIT_PERIOD_WITH_BUYER,
       SINGLE: { TOTAL_CONTRACT_VALUE },
       MULTIPLE: { TOTAL_MONTHS_OF_COVER, TOTAL_SALES_TO_BUYER, MAXIMUM_BUYER_WILL_OWE },
     },
-    ABOUT_GOODS_OR_SERVICES: { DESCRIPTION },
   },
-  YOUR_BUYER: {
-    COMPANY_OR_ORGANISATION: { NAME, REGISTRATION_NUMBER, ADDRESS, FIRST_NAME, LAST_NAME, POSITION, WEBSITE },
-  },
-} = FIELD_IDS.INSURANCE;
+} = INSURANCE_FIELD_IDS;
 
 /**
  * NUMBER_FIELDS
@@ -49,165 +42,6 @@ export const NUMBER_FIELDS = [
   ESTIMATED_ANNUAL_TURNOVER,
   PERCENTAGE_TURNOVER,
 ];
-
-/**
- * STRING_NUMBER_FIELDS
- * Add all string fields here that could be numbers
- * Explicit exemptions list of field IDs in the insurance forms that are string fields that could be submitted as a pure number.
- * If one of these fields is pure numbers, our default "sanitise data" behaviour is to transform a string number into a number type.
- * However, we do not want this to happen with these fields. This means that:
- * - The data is saved correctly
- * - We avoid adding additional error handling in the UI.
- * - We avoid a "problem with service" page scenario where the data save fails, because it tries to save a number type as a text string type.
- * @returns {Array} Field IDs
- */
-export const STRING_NUMBER_FIELDS = [
-  SECURITY_CODE,
-  CREDIT_PERIOD_WITH_BUYER,
-  DESCRIPTION,
-  COMPANY_NUMBER,
-  COMPANY_SIC,
-  PHONE_NUMBER,
-  GOODS_OR_SERVICES,
-  ADDRESS_LINE_1,
-  ADDRESS_LINE_2,
-  POSTCODE,
-  BROKER_NAME,
-  NAME,
-  REGISTRATION_NUMBER,
-  ADDRESS,
-  FIRST_NAME,
-  LAST_NAME,
-  POSITION,
-];
-
-/**
- * shouldChangeToNumber
- * Check if a form field value should change from a string to a number
- * @param {Object} Field key and value
- * @returns {Boolean}
- */
-export const shouldChangeToNumber = ({ key, value }: SanitiseValueObjParams) => {
-  if (key) {
-    if (STRING_NUMBER_FIELDS.includes(key) || isEmptyString(String(value))) {
-      return false;
-    }
-  }
-
-  if (typeof value === 'string') {
-    const stripped = stripCommas(value);
-
-    if (isNumber(stripped)) {
-      return true;
-    }
-  }
-
-  return false;
-};
-
-/**
- * replaceCharactersWithCharacterCode
- * Replace certain characters with character codes
- * @param {String} Field value
- * @returns {String}
- */
-export const replaceCharactersWithCharacterCode = (str: string) =>
-  str
-    .replace(/&/g, '&amp;')
-    .replace(/</g, '&lt;')
-    .replace(/>/g, '&gt;')
-    .replace(/"/g, '&quot;')
-    .replace(/'/g, '&#x27;')
-    .replace(/\//g, '&#x2F;')
-    .replace(/\*/g, '&#42;');
-
-/**
- * sanitiseValue
- * Sanitise a form field value
- * @param {Object} Field key and value
- * @returns {Boolean}
- */
-export const sanitiseValue = ({ key, value }: SanitiseValueObjParams) => {
-  if (value === 'true' || value === true) {
-    return true;
-  }
-
-  if (value === 'false' || value === false) {
-    return false;
-  }
-
-  if (shouldChangeToNumber({ key, value })) {
-    const stripped = stripCommas(String(value));
-
-    return Number(stripped);
-  }
-
-  if (key) {
-    /**
-     * Do not sanitise a valid website address.
-     * Otherwise, the it becomes invalid.
-     */
-    if (key === WEBSITE && isValidWebsiteAddress(String(value))) {
-      return value;
-    }
-  }
-
-  return replaceCharactersWithCharacterCode(String(value));
-};
-
-/**
- * sanitiseObject
- * Sanitise an object
- * - If an object key value is an array, call sanitiseArray.
- * - If an object key value is an object, call sanitiseObject.
- * - Otherwise, sanitiseValue is used for each value in the object.
- * @param {Object}
- * @returns {Boolean}
- */
-export const sanitiseObject = (obj: ObjectType) => {
-  const sanitised = {};
-
-  Object.keys(obj).forEach((key) => {
-    const value = obj[key];
-
-    if (value) {
-      sanitised[key] = sanitiseValue({ key, value });
-
-      if (Array.isArray(value)) {
-        sanitised[key] = sanitiseArray(key, value);
-      } else if (isAnObjectWithKeysAndValues(value)) {
-        sanitised[key] = sanitiseObject(value);
-      }
-    }
-  });
-
-  return sanitised;
-};
-
-/**
- * sanitiseArray
- * Sanitise an array
- * @param {String} Field key
- * @param {Array}
- * @returns {Array}
- */
-export const sanitiseArray = (key: string, arr: Array<string> | Array<object>) => {
-  const sanitised = arr.map((value: string | object) => {
-    const objectWithKeysAndValues = isAnObjectWithKeysAndValues(value);
-
-    if (objectWithKeysAndValues) {
-      return sanitiseObject(objectWithKeysAndValues);
-    }
-
-    if (typeof value === 'string') {
-      return sanitiseValue({ key, value });
-    }
-
-    return null;
-  });
-
-  return sanitised;
-};
 
 /**
  * isDayMonthYearField
@@ -231,13 +65,18 @@ export const isDayMonthYearField = (fieldName: string): boolean => {
  * @returns {Boolean}
  */
 export const shouldIncludeAndSanitiseField = (key: string, value: string) => {
-  // do not include day/month/year fields, these should be captured as timestamps.
+  /**
+   * Do not include day/month/year fields.
+   * These should be captured as timestamps.
+   */
   if (isDayMonthYearField(key)) {
     return false;
   }
 
-  // EMS-793: do not include number fields that are empty.
-  // empty number fields cannot be sent and saved via the API because it is not a number.
+  /**
+   * EMS-793: Do not include number fields that are empty.
+   * Empty number fields cannot be sent and saved via the API because it is not a number.
+   */
   if (NUMBER_FIELDS.includes(key) && isEmptyString(value)) {
     return false;
   }
@@ -254,7 +93,7 @@ export const shouldIncludeAndSanitiseField = (key: string, value: string) => {
  */
 export const sanitiseFormField = (key: string, value: string | boolean | ObjectType | Array<string>) => {
   if (Array.isArray(value)) {
-    return sanitiseArray(key, value);
+    return sanitiseArrayOfStrings(value);
   }
 
   if (isValidDate(value)) {
